intent-recognition-server/requirements.txt
requests
python-dotenv
uvicorn[standard]
fastapi
python-multipart
langchain
OpenAI
langchain-community